+++ distage/distage-testkit-core/src/main/scala/izumi/distage/testkit/model/TestActivationStrategy.scala
@@ -1,0 +1,9 @@
+package izumi.distage.testkit.model
+
+sealed trait TestActivationStrategy
+
+object TestActivationStrategy {
+  case object IgnoreConfig extends TestActivationStrategy
+
+  case class LoadConfig(ignoreUnknown: Boolean, warnUnset: Boolean) extends TestActivationStrategy
+}
