+++ distage/distage-testkit-core/src/main/scala/izumi/distage/testkit/runner/BootstrapFactory.scala
@@ -1,4 +1,4 @@
-package izumi.distage.testkit.services.dstest
+package izumi.distage.testkit.runner
 
 import distage.config.AppConfig
 import izumi.distage.framework.config.PlanningOptions
