--- distage/distage-testkit-core/src/main/scala/izumi/distage/testkit/services/dstest/TestConfiguration.scala
@@ -1,7 +1,0 @@
-package izumi.distage.testkit.services.dstest
-
-import izumi.distage.testkit.TestConfig
-
-trait TestConfiguration {
-  protected def config: TestConfig
-}
