+++ distage/distage-testkit-core/src/main/scala/izumi/distage/testkit/spec/TestConfiguration.scala
@@ -1,0 +1,7 @@
+package izumi.distage.testkit.spec
+
+import izumi.distage.testkit.model.TestConfig
+
+trait TestConfiguration {
+  protected def config: TestConfig
+}
