+++ distage/distage-testkit-scalatest/src/test/scala-2/izumi/distage/testkit/distagesuite/integration/IntegrationTest1Test.scala
@@ -6,7 +6,7 @@ import izumi.distage.model.provisioning.IntegrationCheck
 import izumi.distage.model.definition.{Lifecycle, ModuleDef}
 import izumi.functional.quasi.QuasiIO
 import izumi.distage.modules.{DefaultModule, DefaultModule2}
-import izumi.distage.testkit.TestConfig
+import izumi.distage.testkit.model.TestConfig
 import izumi.distage.testkit.scalatest.{Spec1, Spec2}
 import izumi.functional.bio.catz.*
 import izumi.functional.bio.{Applicative2, ApplicativeError2, F}
